New SDK notifications

+++ includes/Ipn/UpdateBlikAlias.php
@@ -2,36 +2,26 @@
 
 namespace Tpay\Ipn;
 
+use Tpay\OpenApi\Model\Objects\NotificationBody\BlikAliasRegister;
+use Tpay\OpenApi\Model\Objects\NotificationBody\BlikAliasUnregister;
 use Tpay\OpenApi\Utilities\TpayException;
 
 class UpdateBlikAlias implements IpnInterface
 {
-    public function parseNotification($response)
+    public function handle($notification): void
     {
-        [, $uid] = explode('_', $response['msg_value']['value'], 2);
-        switch ($response['event']) {
-            case 'ALIAS_REGISTER':
-                $this->saveBlikAlias($response['msg_value']['value'], $uid);
-                break;
-            case 'ALIAS_UNREGISTER':
-                $this->removeBlikAlias($uid);
-                break;
-            default:
-                throw new TpayException('Unknown event: '.$response['event']);
+        $alias = $notification->value->getValue();
+        [, $uid] = explode('_', $alias, 2);
+
+        if ($notification instanceof BlikAliasRegister) {
+            update_user_meta($uid, 'tpay_alias_blik', $alias);
+            return;
         }
-    }
+        if ($notification instanceof BlikAliasUnregister) {
+            delete_user_meta($uid, 'tpay_alias_blik');
+            return;
+        }
 
-    public function removeBlikAlias($uid)
-    {
-        delete_user_meta($uid, 'tpay_alias_blik');
-        header('HTTP/1.1 200 OK');
-        echo 'TRUE';
-    }
-
-    public function saveBlikAlias($alias, $uid)
-    {
-        update_user_meta($uid, 'tpay_alias_blik', $alias);
-        header('HTTP/1.1 200 OK');
-        echo 'TRUE';
+        throw new TpayException('Invalid notification object for BLIK strategy');
     }
 }
